feat: session send/get

+++ session/manager.py
@@ -1,58 +1,19 @@
 import json
 import os
 import shutil
-import threading
+
+from collections import OrderedDict
 from typing import List
 
 import error
 from rwlock import RWLock
-
-
-class SessionText:
-    def __init__(self, d: str):
-        # 参数
-        try:
-            with open(os.path.join(d, "params.json"), "r") as f:
-                self.params: dict = json.loads(f.read())
-        except FileNotFoundError:
-            raise error.InvalidParamError(f"no such text {d}")
-        except json.JSONDecodeError as e:
-            raise error.InternalError(f"invalid text params.json in {d}: {e}")
-
-        try:
-            # 创建提示
-            with open(os.path.join(d, "create.txt"), "r") as f:
-                self.t_create: str = f.read()
-
-            # 压缩需要的提示
-            with open(os.path.join(d, "compress.txt"), "r") as f:
-                self.t_compress: str = f.read()
-
-            # 继承需要的提示
-            with open(os.path.join(d, "inherit.txt"), "r") as f:
-                self.t_inherit: str = f.read()
-        except OSError as e:
-            raise error.InternalError(f"invalid text in {d}: {e}")
-
-
-class Session:
-    def __init__(self, d: str):
-        with open(os.path.join(d, "index.json"), "r") as f:
-            j = json.loads(f.read())
-        self.id: str = j["id"]
-        self.type: str = j["type"]
-        self.params: dict = j["params"]
-
-    def __dict__(self) -> dict:
-        return {
-            "id": self.id,
-            "type": self.type,
-            "params": self.params,
-        }
+from schedule import Scheduler
+from session.session import Session
+from session.text import SessionText
 
 
 class SessionManager:
-    def __init__(self, text: str, database: str):
+    def __init__(self, text: str, database: str, scheduler: Scheduler):
         assert len(text) > 0
         assert len(database) > 0
 
@@ -60,23 +21,50 @@ class SessionManager:
 
         self.__text_path = text
         self.__database = database
+        self.__scheduler = scheduler
 
-    def list(self) -> List[Session]:
-        self.__lock.acquire_read()
+        self.__texts: OrderedDict[str, SessionText] = OrderedDict()
+        self.__load_text()
+
+        self.__sessions: OrderedDict[str, Session] = OrderedDict()
+        self.__load_sessions()
+
+    def __load_text(self):
         try:
-            result = []
+            for type_ in os.listdir(self.__text_path):
+                text_path = os.path.join(self.__text_path, type_)
+                if not os.path.isdir(text_path):
+                    continue
+                try:
+                    t = SessionText(text_path)
+                except (FileNotFoundError, json.JSONDecodeError, KeyError) as e:
+                    print(e)
+                    continue
+                self.__texts[type_] = t
+        except OSError as e:
+            raise error.InternalError(e)
+
+    def __load_sessions(self):
+        try:
             for id_ in os.listdir(self.__database):
                 session_path = os.path.join(self.__database, id_)
                 if not os.path.isdir(session_path):
                     continue
                 try:
-                    s = Session(session_path)
+                    s = Session(session_path, self.__texts, self.__scheduler)
                 except (FileNotFoundError, json.JSONDecodeError, KeyError) as e:
                     print(e)
                     continue
-                result.append(s)
+                self.__sessions[id_] = s
         except OSError as e:
             raise error.InternalError(e)
+
+    def list(self) -> List[Session]:
+        self.__lock.acquire_read()
+        try:
+            result = []
+            for id_ in self.__sessions:
+                result.append(self.__sessions[id_])
         finally:
             self.__lock.release_read()
         return result
@@ -86,6 +74,8 @@ class SessionManager:
             raise error.InvalidParamError(f"invalid session id: {id_}")
         if len(type_.strip()) == 0:
             raise error.InvalidParamError(f"invalid session type: {type_}")
+        if id_ in self.__sessions:
+            raise error.InvalidParamError(f"session already exists: {id_}")
 
         self.__lock.acquire()
         try:
@@ -104,8 +94,10 @@ class SessionManager:
                     "id": id_,
                     "type": type_,
                     "params": params,
-                }))
-            return Session(os.path.join(self.__database, id_))
+                }, ensure_ascii=False, indent=2))
+            s = Session(os.path.join(self.__database, id_), self.__texts, self.__scheduler)
+            self.__sessions[id_] = s
+            return s
         except FileExistsError:
             raise error.InvalidParamError(f"session already exists: {id_}")
         except OSError as e:
@@ -113,12 +105,27 @@ class SessionManager:
         finally:
             self.__lock.release()
 
+    def get(self, id_: str) -> Session:
+        if len(id_.strip()) == 0:
+            raise error.InvalidParamError(f"invalid session id: {id_}")
+        s = self.__sessions.get(id_)
+        if s is None:
+            raise error.InvalidParamError(f"no such session: {id_}")
+        return s
+
+    def get_default(self) -> Session:
+        if len(self.__sessions) > 0:
+            return self.__sessions[next(iter(self.__sessions))]
+
     def remove(self, id_: str):
         if len(id_.strip()) == 0:
             raise error.InvalidParamError(f"invalid session id: {id_}")
+        if id_ not in self.__sessions:
+            raise error.InvalidParamError(f"no such session: {id_}")
 
         self.__lock.acquire()
         try:
+            del self.__sessions[id_]
             shutil.rmtree(os.path.join(self.__database, id_))
         except FileNotFoundError:
             raise error.InvalidParamError(f"no such session: {id_}")
